refactor: extract budget details sidebar
